Move modules related to transactions to transactor module
