+++ chains/evm/transactor/monitored/monitored_test.go
@@ -8,10 +8,10 @@ import (
 	"time"
 
 	mock_calls "github.com/ChainSafe/sygma-core/chains/evm/mock"
-	"github.com/ChainSafe/sygma-core/chains/evm/transaction"
 	"github.com/ChainSafe/sygma-core/chains/evm/transactor"
 	mock_transactor "github.com/ChainSafe/sygma-core/chains/evm/transactor/mock"
 	"github.com/ChainSafe/sygma-core/chains/evm/transactor/monitored"
+	"github.com/ChainSafe/sygma-core/chains/evm/transactor/transaction"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/golang/mock/gomock"
